+++ crates/matrix-sdk-base/src/response_processors/state_events.rs
@@ -62,11 +62,11 @@ pub mod sync {
         }))
     }
 
-    /// Dispatch the state events and return the new users for this room.
+    /// Dispatch the sync state events and return the new users for this room.
     ///
-    /// `raw_events` and `events` must be generated from [`collect_sync`].
+    /// `raw_events` and `events` must be generated from [`collect`].
     /// Events must be exactly the same list of events that are in
-    /// raw_events, but deserialised. We demand them here to avoid
+    /// `raw_events`, but deserialised. We demand them here to avoid
     /// deserialising multiple times.
     #[instrument(skip_all, fields(room_id = ?room_info.room_id))]
     pub async fn dispatch_and_get_new_users(
@@ -115,9 +115,20 @@ pub mod sync {
 
 /// Collect [`AnyStrippedStateEvent`].
 pub mod stripped {
-    use ruma::events::AnyStrippedStateEvent;
+    use std::{collections::BTreeMap, iter};
 
-    use super::{Context, Raw};
+    use ruma::{
+        events::AnyStrippedStateEvent,
+        push::{Action, Ruleset},
+        OwnedRoomId,
+    };
+    use tracing::instrument;
+
+    use super::{super::timeline, Context, Raw};
+    use crate::{
+        deserialized_responses::RawAnySyncOrStrippedTimelineEvent, store::BaseStateStore,
+        sync::Notification, Result, Room, RoomInfo,
+    };
 
     /// Collect [`AnyStrippedStateEvent`] to [`AnyStrippedStateEvent`].
     pub fn collect(
@@ -125,6 +136,74 @@ pub mod stripped {
         raw_events: &[Raw<AnyStrippedStateEvent>],
     ) -> (Vec<Raw<AnyStrippedStateEvent>>, Vec<AnyStrippedStateEvent>) {
         super::collect(raw_events)
+    }
+
+    /// Dispatch the stripped state events.
+    ///
+    /// `raw_events` and `events` must be generated from [`collect`].
+    /// Events must be exactly the same list of events that are in
+    /// `raw_events`, but deserialised. We demand them here to avoid
+    /// deserialising multiple times.
+    ///
+    /// Dispatch the stripped state events in `invite_state` or `knock_state`,
+    /// modifying the room's info and posting notifications as needed.
+    ///
+    /// * `raw_events` and `events` - The contents of `invite_state` in the form
+    ///   of list of pairs of raw stripped state events with their deserialized
+    ///   counterpart.
+    /// * `room` - The [`Room`] to modify.
+    /// * `room_info` - The current room's info.
+    /// * `push_rules` - The push rules for this room.
+    /// * `changes` - The accumulated list of changes to apply once the
+    ///   processing is finished.
+    /// * `notifications` - Notifications to post for the current room.
+    /// * `state_store` — The state store.
+    #[instrument(skip_all, fields(room_id = ?room_info.room_id))]
+    pub(crate) async fn dispatch_invite_or_knock(
+        context: &mut Context,
+        (raw_events, events): (&[Raw<AnyStrippedStateEvent>], &[AnyStrippedStateEvent]),
+        room: &Room,
+        room_info: &mut RoomInfo,
+        push_rules: &Ruleset,
+        notifications: &mut BTreeMap<OwnedRoomId, Vec<Notification>>,
+        state_store: &BaseStateStore,
+    ) -> Result<()> {
+        let mut state_events = BTreeMap::new();
+
+        for (raw_event, event) in iter::zip(raw_events, events) {
+            room_info.handle_stripped_state_event(event);
+            state_events
+                .entry(event.event_type())
+                .or_insert_with(BTreeMap::new)
+                .insert(event.state_key().to_owned(), raw_event.clone());
+        }
+
+        context
+            .state_changes
+            .stripped_state
+            .insert(room_info.room_id().to_owned(), state_events.clone());
+
+        // We need to check for notifications after we have handled all state
+        // events, to make sure we have the full push context.
+        if let Some(push_context) =
+            timeline::get_push_room_context(context, room, room_info, state_store).await?
+        {
+            // Check every event again for notification.
+            for event in state_events.values().flat_map(|map| map.values()) {
+                let actions = push_rules.get_actions(event, &push_context);
+
+                if actions.iter().any(Action::should_notify) {
+                    notifications.entry(room.room_id().to_owned()).or_default().push(
+                        Notification {
+                            actions: actions.to_owned(),
+                            event: RawAnySyncOrStrippedTimelineEvent::Stripped(event.clone()),
+                        },
+                    );
+                }
+            }
+        }
+
+        Ok(())
     }
 }
